docs: simplify DeltaSelect explanation
