+++ docs/agent-layer/CONTEXT.md
@@ -37,7 +37,7 @@ Do not duplicate information that belongs in other memory files:
 
 ## DeepSWE benchmark
 
-- The public method and website brand is **DeltaSelect**. Use `DeltaSelect` for the top-level page, navigation, metadata, and explanatory copy; DeepSWE is the first public evidence dataset and may remain in the established `/deepswe-planner` implementation route. DeltaSelect is planned for v0.18.0 alongside its companion arXiv paper; the paper is not public yet.
+- The public method and website brand is **DeltaSelect**. Use `DeltaSelect` for the top-level page, navigation, metadata, and explanatory copy. The interactive page is `/deltaselect-tool`; DeepSWE is the first public evidence dataset, and its static implementation assets remain under `site/static/deepswe-planner/`. DeltaSelect is planned for v0.18.0 alongside its companion arXiv paper; the paper is not public yet.
 - The sole public benchmark workflow is `al benchmark run <study.toml>` plus developer-facing `benchmark readiness`. The study names a v1/v2 website selection and explicit content-addressed experiments; task filters and worker count are invocation controls only. Private legacy campaign evidence under `.agent-layer/state/benchmarks/deepswe/campaigns/` remains untouched and is not migrated.
 - The website exports schema-v2 `deepswe-benchmark-selection` with explicit manual exclusions. It ranks deterministic correlation evidence, excludes nonpositive OLS residual variances, applies exclusions and published-model headroom before budget walking, and renormalizes remaining precision weights. Its optional published comparison is descriptive and inert to selection.
 - The embedded public v1.1 snapshot is a reviewed offline input, not runtime web data. Its results use `mini-swe-agent`, so published-only output is reference-relative and cannot attribute a difference to Agent Layer.
