+++ site/pages/deltaselect.mdx
@@ -14,10 +14,10 @@ import Link from '@docusaurus/Link';
 
 # DeltaSelect
 
-DeltaSelect is a benchmark task-selection method for teams that need useful experimental evidence without paying to rerun an entire software-engineering benchmark. It ranks tasks by how reliably they track the full benchmark, accounts for uncertainty and available improvement headroom, and selects the strongest set that fits a fixed budget.
+DeltaSelect helps any developer choose a smaller set of benchmark tasks when running the full benchmark costs too much. It uses results from earlier runs to find tasks that tend to move with the full benchmark, then selects the most useful set that fits your budget. You can see the estimated cost, uncertainty, and evidence behind every choice.
 
 <div className="hero-buttons">
-  <Link to="/deepswe-planner" className="hero-button hero-button--primary">Open DeltaSelect</Link>
+  <Link to="/deltaselect-tool" className="hero-button hero-button--primary">Open DeltaSelect</Link>
   <Link to="/docs/reference#benchmark-runner" className="hero-button hero-button--secondary">Benchmark runner</Link>
 </div>
 
@@ -27,25 +27,25 @@ DeltaSelect debuts in Agent Layer v0.18.0. A companion arXiv paper describing th
 
 {/* Add the companion arXiv paper URL here when it is published. */}
 
-## Why benchmark selection matters
+## Why select benchmark tasks?
 
-Evaluating an instruction, skill, harness, or model change across every task and repeated run of a major coding benchmark can cost far more than the change itself. A tiny arbitrary subset is cheaper, but it may be noisy, saturated, or unrelated to movement in the complete benchmark.
+Running every task in a coding benchmark can be expensive, especially when each task needs several runs. Picking a few tasks at random is cheaper, but the result may be noisy or may not reflect the full benchmark.
 
-DeltaSelect turns that tradeoff into an explicit experiment-design problem:
+DeltaSelect gives you a more deliberate way to choose:
 
-- **Spend only the budget you set.** Tasks are added in evidence-ranked order while expected published-run cost remains within the selected budget.
-- **Prefer informative tasks.** Ranking uses the conservative end of each task's estimated correlation with the complete benchmark rather than a single optimistic point estimate.
-- **Preserve a fixed estimator.** Selected tasks carry calibrated weights so the resulting score is defined before the experiment runs.
-- **Expose exclusions and uncertainty.** Manual exclusions, incomplete evidence, expected score, expected cost, and uncertainty remain visible instead of being hidden behind one recommendation.
-- **Export a reproducible study.** The browser produces a content-addressed selection document and a study template for `al benchmark run`.
+- **Set a budget.** DeltaSelect stops adding tasks before the estimated run cost exceeds the amount you choose.
+- **Choose tasks backed by prior results.** It prefers tasks that have tracked the full benchmark consistently, while accounting for uncertainty and room for improvement.
+- **Define the score before you run.** Each selected task gets a fixed weight, so the scoring method cannot change after results arrive.
+- **See the tradeoffs.** The tool shows excluded tasks, missing data, estimated cost, expected score, and uncertainty.
+- **Repeat the study.** Export the selection and a `study.toml` template for `al benchmark run`.
 
-## What the current tool does
+## Try DeltaSelect
 
-The first public DeltaSelect application uses a reviewed offline snapshot of repeated [DeepSWE](https://deepswe.datacurve.ai/blog/deepswe) runs. Choose a model, reasoning level, budget, repetition count, and optional minimum headroom. DeltaSelect then shows the selected tasks, normalized weights, expected published score and cost, and the evidence behind every ranking.
+The interactive tool uses a reviewed snapshot of repeated [DeepSWE](https://deepswe.datacurve.ai/blog/deepswe) runs. Choose a model, reasoning level, budget, number of runs, and optional minimum room for improvement. DeltaSelect shows the selected tasks, their weights, the estimated score and cost, and the data behind each choice.
 
-An optional comparison configuration estimates how that second published configuration would perform on the same fixed tasks, repetitions, and weights. It is descriptive only and never changes the selection.
+You can also compare a second model configuration against the same tasks and weights. The comparison does not change which tasks DeltaSelect selects.
 
-## From selection to a real experiment
+## Run the study
 
 DeltaSelect connects directly to Agent Layer's [benchmark runner](/docs/reference#benchmark-runner):
 
@@ -54,10 +54,10 @@ DeltaSelect connects directly to Agent Layer's [benchmark runner](/docs/referenc
 3. Run `al benchmark run <study.toml>` to execute missing cells and preserve immutable evidence.
 4. Compare score, cost, and workflow-conformance results in the generated report.
 
-This makes it practical to test changes such as an always-loaded instruction revision, a new Agent Skill, or an Agent Dispatch workflow against the same declared task set.
+Use the same declared task set to test an instruction change, a new Agent Skill, an Agent Dispatch workflow, or another coding-agent improvement.
 
 ## Scope and interpretation
 
-DeltaSelect optimizes evidence for a declared selection and budget. A result applies to those selected tasks; it is not automatically representative of every task in the source benchmark. Historical detectability also cannot predict the effect size of a new treatment. The tool keeps those boundaries explicit so teams can make a proportionate evaluation decision instead of treating a small benchmark as ground truth.
+DeltaSelect does not make a small task set equivalent to the full benchmark. A result describes the selected tasks, and past benchmark behavior cannot predict exactly how a new change will perform. Use the cost and uncertainty shown by the tool when deciding how much confidence to place in the result.
 
-<Link to="/deepswe-planner">Open the interactive DeltaSelect tool →</Link>
+<Link to="/deltaselect-tool">Open the interactive DeltaSelect tool →</Link>
